+++ br/pkg/redact/redact_test.go
@@ -14,10 +14,10 @@ func TestRedact(t *testing.T) {
 	redacted, secret := "?", "secret"
 
 	redact.InitRedact(false)
-	require.Equal(t, redact.String(secret), secret)
+	require.Equal(t, redact.Value(secret), secret)
 	require.Equal(t, redact.Key([]byte(secret)), hex.EncodeToString([]byte(secret)))
 
 	redact.InitRedact(true)
-	require.Equal(t, redact.String(secret), redacted)
+	require.Equal(t, redact.Value(secret), redacted)
 	require.Equal(t, redact.Key([]byte(secret)), redacted)
 }
