+++ pkg/util/dbutil/table_test.go
@@ -12,14 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dbutil
+package dbutil_test
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/pingcap/tidb/pkg/parser"
+	"github.com/pingcap/tidb/pkg/parser/model"
 	"github.com/pingcap/tidb/pkg/parser/mysql"
+	_ "github.com/pingcap/tidb/pkg/planner/core" // to setup expression.EvalAstExpr. See: https://github.com/pingcap/tidb/blob/a94cff903cd1e7f3b050db782da84273ef5592f4/planner/core/optimizer.go#L202
 	"github.com/pingcap/tidb/pkg/types"
+	"github.com/pingcap/tidb/pkg/util/dbutil"
+	"github.com/pingcap/tidb/pkg/util/dbutil/dbutiltest"
 	"github.com/pingcap/tidb/pkg/util/schemacmp"
 	"github.com/stretchr/testify/require"
 )
@@ -114,7 +119,7 @@ func TestTable(t *testing.T) {
 	}
 
 	for _, testCase := range testCases {
-		tableInfo, err := GetTableInfoBySQL(testCase.sql, parser.New())
+		tableInfo, err := dbutiltest.GetTableInfoBySQL(testCase.sql, parser.New())
 		require.NoError(t, err)
 		for i, column := range tableInfo.Columns {
 			require.Equal(t, column.Name.O, testCase.columns[i])
@@ -128,24 +133,69 @@ func TestTable(t *testing.T) {
 			}
 		}
 
-		col := FindColumnByName(tableInfo.Columns, testCase.colName)
+		col := dbutil.FindColumnByName(tableInfo.Columns, testCase.colName)
 		require.Equal(t, col != nil, testCase.fineCol)
 	}
 }
 
+// EqualTableInfo returns true if this two table info have same columns and indices
+func EqualTableInfo(tableInfo1, tableInfo2 *model.TableInfo) (bool, string) {
+	// check columns
+	if len(tableInfo1.Columns) != len(tableInfo2.Columns) {
+		return false, fmt.Sprintf("column num not equal, one is %d another is %d", len(tableInfo1.Columns), len(tableInfo2.Columns))
+	}
+
+	for j, col := range tableInfo1.Columns {
+		if col.Name.O != tableInfo2.Columns[j].Name.O {
+			return false, fmt.Sprintf("column name not equal, one is %s another is %s", col.Name.O, tableInfo2.Columns[j].Name.O)
+		}
+		if col.GetType() != tableInfo2.Columns[j].GetType() {
+			return false, fmt.Sprintf("column %s's type not equal, one is %v another is %v", col.Name.O, col.GetType(), tableInfo2.Columns[j].GetType())
+		}
+	}
+
+	// check index
+	if len(tableInfo1.Indices) != len(tableInfo2.Indices) {
+		return false, fmt.Sprintf("index num not equal, one is %d another is %d", len(tableInfo1.Indices), len(tableInfo2.Indices))
+	}
+
+	index2Map := make(map[string]*model.IndexInfo)
+	for _, index := range tableInfo2.Indices {
+		index2Map[index.Name.O] = index
+	}
+
+	for _, index1 := range tableInfo1.Indices {
+		index2, ok := index2Map[index1.Name.O]
+		if !ok {
+			return false, fmt.Sprintf("index %s not exists", index1.Name.O)
+		}
+
+		if len(index1.Columns) != len(index2.Columns) {
+			return false, fmt.Sprintf("index %s's columns num not equal, one is %d another is %d", index1.Name.O, len(index1.Columns), len(index2.Columns))
+		}
+		for j, col := range index1.Columns {
+			if col.Name.O != index2.Columns[j].Name.O {
+				return false, fmt.Sprintf("index %s's column not equal, one has %s another has %s", index1.Name.O, col.Name.O, index2.Columns[j].Name.O)
+			}
+		}
+	}
+
+	return true, ""
+}
+
 func TestTableStructEqual(t *testing.T) {
 	createTableSQL1 := "CREATE TABLE `test`.`atest` (`id` int(24), `name` varchar(24), `birthday` datetime, `update_time` time, `money` decimal(20,2), primary key(`id`))"
-	tableInfo1, err := GetTableInfoBySQL(createTableSQL1, parser.New())
+	tableInfo1, err := dbutiltest.GetTableInfoBySQL(createTableSQL1, parser.New())
 	require.NoError(t, err)
 
 	createTableSQL2 := "CREATE TABLE `test`.`atest` (`id` int(24) NOT NULL, `name` varchar(24), `birthday` datetime, `update_time` time, `money` decimal(20,2), primary key(`id`))"
-	tableInfo2, err := GetTableInfoBySQL(createTableSQL2, parser.New())
+	tableInfo2, err := dbutiltest.GetTableInfoBySQL(createTableSQL2, parser.New())
 	require.NoError(t, err)
 
 	createTableSQL3 := `CREATE TABLE "test"."atest" ("id" int(24), "name" varchar(24), "birthday" datetime, "update_time" time, "money" decimal(20,2), unique key("id"))`
 	p := parser.New()
 	p.SetSQLMode(mysql.ModeANSIQuotes)
-	tableInfo3, err := GetTableInfoBySQL(createTableSQL3, p)
+	tableInfo3, err := dbutiltest.GetTableInfoBySQL(createTableSQL3, p)
 	require.NoError(t, err)
 
 	equal, _ := EqualTableInfo(tableInfo1, tableInfo2)
@@ -157,7 +207,7 @@ func TestTableStructEqual(t *testing.T) {
 
 func TestSchemacmpEncode(t *testing.T) {
 	createTableSQL := "CREATE TABLE `test`.`atest` (`id` int(24), primary key(`id`))"
-	tableInfo, err := GetTableInfoBySQL(createTableSQL, parser.New())
+	tableInfo, err := dbutiltest.GetTableInfoBySQL(createTableSQL, parser.New())
 	require.NoError(t, err)
 
 	table := schemacmp.Encode(tableInfo)
